+++ packages/typespec-client-generator-core/src/decorators.ts
@@ -166,7 +166,7 @@ export const $client: ClientDecorator = (
   const service =
     explicitService?.kind === "Namespace"
       ? explicitService
-      : (findClientService(context.program, target) ?? (target as any));
+      : (findClientService(context.program, target, scope) ?? (target as any));
   if (!name.endsWith("Client")) {
     reportDiagnostic(context.program, {
       code: "client-name",
@@ -196,7 +196,8 @@ export const $client: ClientDecorator = (
 
 function findClientService(
   program: Program,
-  client: Namespace | Interface
+  client: Namespace | Interface,
+  scope?: LanguageScopes
 ): Namespace | Interface | undefined {
   let current: Namespace | undefined = client as any;
   while (current) {
@@ -205,8 +206,8 @@ function findClientService(
       return current;
     }
     const client = program.stateMap(clientKey).get(current);
-    if (client && client[AllScopes]) {
-      return client[AllScopes].service;
+    if (client && client[scope ?? AllScopes]) {
+      return client[scope ?? AllScopes].service;
     }
     current = current.namespace;
   }
@@ -359,7 +360,7 @@ export const $operationGroup: OperationGroupDecorator = (
     });
     return;
   }
-  const service = findClientService(context.program, target) ?? (target as any);
+  const service = findClientService(context.program, target, scope) ?? (target as any);
   if (!isService(context.program, service)) {
     reportDiagnostic(context.program, {
       code: "client-service",
@@ -458,7 +459,7 @@ export function getOperationGroup(
   type: Namespace | Interface
 ): SdkOperationGroup | undefined {
   let operationGroup: SdkOperationGroup | undefined;
-  const service = findClientService(context.program, type) ?? (type as any);
+  const service = findClientService(context.program, type, context.emitterName) ?? (type as any);
   if (!isService(context.program, service)) {
     reportDiagnostic(context.program, {
       code: "client-service",
@@ -658,7 +659,7 @@ export async function createSdkContext<
     packageName: context.options["package-name"],
     flattenUnionAsEnum: context.options["flatten-union-as-enum"] ?? true,
     apiVersion: options?.versioning?.strategy === "ignore" ? "all" : context.options["api-version"],
-    examplesDirectory: context.options["examples-directory"],
+    examplesDir: context.options["examples-dir"] ?? context.options["examples-directory"],
     decoratorsAllowList: [...defaultDecoratorsAllowList, ...(options?.additionalDecorators ?? [])],
     previewStringRegex: options?.versioning?.previewStringRegex || tcgcContext.previewStringRegex,
   };
@@ -1105,12 +1106,12 @@ function compareModelProperties(modelPropA: ModelProperty, modelPropB: ModelProp
   );
 }
 
-export function $override(
+export const $override = (
   context: DecoratorContext,
   original: Operation,
   override: Operation,
   scope?: LanguageScopes
-) {
+) => {
   // Extract and sort parameter names
   const originalParams = collectParams(original.parameters.properties).sort((a, b) =>
     a.name.localeCompare(b.name)
@@ -1136,7 +1137,7 @@ export function $override(
     });
   }
   setScopedDecoratorData(context, $override, overrideKey, original, override, scope); // eslint-disable-line deprecation/deprecation
-}
+};
 
 /**
  * Gets additional information on how to serialize / deserialize TYPESPEC standard types depending
@@ -1152,3 +1153,9 @@ export function getOverriddenClientMethod(
 ): Operation | undefined {
   return getScopedDecoratorData(context, overrideKey, entity);
 }
+
+export const $useSystemTextJsonConverter: DecoratorFunction = (
+  context: DecoratorContext,
+  entity: Model,
+  scope?: LanguageScopes
+) => {};