+++ packages/typespec-client-generator-core/test/types/built-in-types.test.ts
@@ -19,7 +19,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runner.compileWithBuiltInService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         prop: string;
       }
@@ -35,7 +34,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runner.compileWithBuiltInService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         prop: boolean;
       }
@@ -63,7 +61,6 @@ describe("typespec-client-generator-core: built-in types", () => {
       await runner.compileWithBuiltInService(
         `
         @usage(Usage.input | Usage.output)
-        @access(Access.public)
         model Test {
           prop: ${type};
         }
@@ -82,7 +79,6 @@ describe("typespec-client-generator-core: built-in types", () => {
       await runner.compileWithBuiltInService(
         `
         @usage(Usage.input | Usage.output)
-        @access(Access.public)
         model Test {
           prop: ${type};
         }
@@ -99,7 +95,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runner.compileWithBuiltInService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         prop: decimal;
       }
@@ -115,7 +110,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runner.compileWithBuiltInService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         prop: decimal128;
       }
@@ -131,7 +125,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runner.compileWithBuiltInService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         prop: unknown;
       }
@@ -145,7 +138,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runner.compileWithBuiltInService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         prop: bytes;
       }
@@ -161,7 +153,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runner.compileWithBuiltInService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         @encode(BytesKnownEncoding.base64)
         prop: bytes;
@@ -178,7 +169,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runner.compileWithBuiltInService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         @encode(BytesKnownEncoding.base64url)
         prop: bytes;
@@ -198,7 +188,6 @@ describe("typespec-client-generator-core: built-in types", () => {
       scalar Base64UrlBytes extends bytes;
 
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         value: Base64UrlBytes[];
       }
@@ -224,7 +213,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runnerWithCore.compileWithBuiltInAzureCoreService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         id: Azure.Core.armResourceIdentifier<[
           {
@@ -251,7 +239,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runnerWithCore.compileWithBuiltInAzureCoreService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         urlScalar: url;
 
@@ -313,7 +300,6 @@ describe("typespec-client-generator-core: built-in types", () => {
       scalar Derived extends Base;
 
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         prop: Derived;
       }
@@ -338,7 +324,6 @@ describe("typespec-client-generator-core: built-in types", () => {
     await runner.compileWithBuiltInService(
       `
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         @format("unknown")
         unknownProp: string;
@@ -402,7 +387,6 @@ describe("typespec-client-generator-core: built-in types", () => {
       scalar TestScalar extends string;
       
       @usage(Usage.input | Usage.output)
-      @access(Access.public)
       model Test {
         prop: TestScalar;
       }
@@ -416,5 +400,39 @@ describe("typespec-client-generator-core: built-in types", () => {
     strictEqual(type.description, "title");
     strictEqual(type.details, "doc");
     strictEqual(type.crossLanguageDefinitionId, "TestService.TestScalar");
+  });
+
+  it("integer model property encoded as string", async function () {
+    await runner.compileWithBuiltInService(
+      `
+      @usage(Usage.input | Usage.output)
+      model Test {
+        @encode(string)
+        value: safeint;
+      }
+      `
+    );
+    const sdkType = getSdkTypeHelper(runner);
+    strictEqual(sdkType.kind, "safeint");
+    strictEqual(sdkType.encode, "string");
+    strictEqual(sdkType.baseType, undefined);
+  });
+
+  it("integer scalar encoded as string", async function () {
+    await runner.compileWithBuiltInService(
+      `
+      @encode(string)
+      scalar int32EncodedAsString extends int32;
+
+      @usage(Usage.input | Usage.output)
+      model Test {
+        value: int32EncodedAsString;
+      }
+      `
+    );
+    const sdkType = getSdkTypeHelper(runner);
+    strictEqual(sdkType.kind, "int32");
+    strictEqual(sdkType.encode, "string");
+    strictEqual(sdkType.baseType?.kind, "int32");
   });
 });